Fixed the root of the text clipping issue, and custom CSS
I'll be sure to look harder next time before pushing out a fix that
breaks formatting as much as the last one did. As for custom CSS, I was
able to pull more from the Chrome extension source to get that working.
The Firefox request handling was quite foreign to Safari.